+++ com.unity.ml-agents/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Major Changes
 
 ### Minor Changes
-
+ - Format of console output has changed slightly and now matches the name of the model/summary directory. (#3630, #3616)
 
 ## [0.15.0-preview] - 2020-03-18
 ### Major Changes